+++ CHANGELOG.md
@@ -301,6 +301,9 @@ how to upgrade.
 * The `bundleVersion` property is available on `BugsnagConfiguration` allowing overriding the default plist value.
   [#550](https://github.com/bugsnag/bugsnag-cocoa/pull/550)
 
+* Deliver each event in a separate request to avoid exceeding payload size limit
+  [#424](https://github.com/bugsnag/bugsnag-cocoa/pull/424)
+
 ## 5.23.2 (2020-05-13)
 
 ## Bug Fixes